change variable type to double

+++ app/src/main/java/com/example/fishmarket/data/repository/transaction/source/local/entity/DetailTransactionEntity.kt
@@ -9,6 +9,6 @@ data class DetailTransactionEntity(
     @PrimaryKey @ColumnInfo(name = "id") var id: String,
     @ColumnInfo(name = "id_transaction") var id_transaction:String,
     @ColumnInfo(name = "id_menu") var id_menu: String,
-    @ColumnInfo(name = "quantity") var quantity: Int,
+    @ColumnInfo(name = "quantity") var quantity: Double,
     @ColumnInfo(name = "price") var price: Int
 )
