+++ app/src/main/java/com/example/fishmarket/utilis/Product.kt
@@ -1,3 +1,3 @@
 package com.example.fishmarket.utilis
 
-data class Product(var id: String, var name: String, var price: Int, var quantity: Int)
+data class Product(var id: String, var name: String, var price: Int, var quantity: Double)
